Fix MismatchingStateError when state is a bytes object
Since we use JWT to store the state/session it can sometimes be stored as a bytes object instead of a string. The Oauth2 lib expects state to be a string. This detects if state is a bytes object and if so, decodes it to a string to pass to oauth2.

I came across this issue while testing a real deployment.